refactor(auth): extract auth errors and telemetry
